Incompatible change: Hide by default
This changes the default behavior to hide all chats except those that
are not set to be hidden. Unfortunately with the boolean setting we
cannot distguish between "not set" and "show chat". Therefore we need to
set a new setting.

That also means that the old settings will be lost in future versions.

Signed-off-by: Konrad Gräfe <konradgraefe@aol.com>